requirements.txt
streamlit 
moviepy 
google-cloud-speech 
google-generativeai 
sentence-transformers 
python-dotenv
numpy==1.26.4
SpeechRecognition
pydub
openai
transformers 
langchain
pinecone-client
langchain-community
pinecone-notebooks
pinecone-text
python-dotenv
sentence_transformers
google-generativeai
pydub